Split meta serialization into its own module.
To avoid circular imports, we have to pass the serializer for DocCursors down.
